requirements.txt
alembic==1.7.6
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
autopage==0.5.0
backcall==0.2.0
beautifulsoup4==4.10.0
black==22.1.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
cliff==3.10.0
cmaes==0.8.2
cmd2==2.3.3
colorlog==6.6.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
filelock==3.6.0
flake8==4.0.1
greenlet==1.1.2
huggingface-hub==0.4.0
idna==3.3
importlib-metadata==4.2.0
importlib-resources==5.4.0
ipykernel==6.9.1
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.10.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
Mako==1.1.6
markdown-it-py==2.0.1
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
mccabe==0.6.1
mdformat==0.7.13
mdurl==0.1.0
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
notebook==6.4.9
numpy==1.21.5
optuna==2.10.0
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pbr==5.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.0
prettytable==3.1.1
prometheus-client==0.13.1
prompt-toolkit==3.0.28
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.11.2
pyparsing==3.0.7
pyperclip==1.8.2
pyrsistent==0.18.1
pysentiment==0.2
python-dateutil==2.8.2
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
regex==2022.1.18
requests==2.27.1
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.31
stack-data==0.2.0
stevedore==3.5.0
terminado==0.13.3
testpath==0.6.0
threadpoolctl==3.1.0
tokenize-rt==4.2.1
tokenizers==0.11.5
tomli==2.0.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.16.2
typed-ast==1.5.2
typing_extensions==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
zipp==3.7.0
